+++ lib/language_pack/base.rb
@@ -153,7 +153,7 @@ private ##################################
     %x{ #{command} 2>/dev/null }
   end
 
-  # run a shell command and stream the ouput
+  # run a shell command and stream the output
   # @param [String] command to be run
   def pipe(command)
     output = ""
@@ -222,5 +222,11 @@ private ##################################
     system("cp -a #{from}/. #{to}")
   end
 
+  # check if the cache content exists
+  # @param [String] relative path of the cache contents
+  # @param [Boolean] true if the path exists in the cache and false if otherwise
+  def cache_exists?(path)
+    File.exists?(cache_base + path)
+  end
 end